requirements.txt
# Core Backend
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4

# LangChain & Agents
langchain==0.1.16
langgraph==0.0.40
langchain-community==0.0.32
langchain-openai==0.1.6

# Vector DB
chromadb==0.4.24
sentence-transformers==2.6.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# DevOps & Cloud
boto3==1.34.84
python-dotenv==1.0.1
requests==2.31.0

# Utilities
loguru==0.7.2
tenacity==8.2.3
